Completely rewrite library
I decided not to split this into separate commits, since this is pretty much a full rewrite,
The API is pretty much the same, but kings were added, so this can theoretically be used to
simulate full games, though ending conditions are not yet implemented. Also, tests were added
and most of the library now has documentation.